cleared errors on address in edit page of both kitchen&org,also changed the display of kitchen selcted menu
